+++ crates/gateway/src/lib.rs
@@ -6,6 +6,6 @@ pub mod rpc_state_reader;
 pub mod starknet_api_test_utils;
 pub mod state_reader;
 pub mod state_reader_test_utils;
-pub mod stateful_transaction_validator;
-pub mod stateless_transaction_validator;
+pub mod stateful_tx_validator;
+pub mod stateless_tx_validator;
 pub mod utils;
